feat: state machine is use case.
