+++ mobile/app/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/mobile/App.kt
@@ -19,7 +19,6 @@ import co.kr.parkjonghun.whatishedoingwithandroid.feature.top.topScreenRoute
 import co.kr.parkjonghun.whatishedoingwithandroid.inside.di.repositoryModule
 import co.kr.parkjonghun.whatishedoingwithandroid.mobile.navigation.AppNavigationState
 import co.kr.parkjonghun.whatishedoingwithandroid.mobile.navigation.rememberAppNavigationState
-import co.kr.parkjonghun.whatishedoingwithandroid.service.statemachine.di.stateMachineModule
 import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.di.useCaseModule
 import co.kr.parkjonghun.whatishedoingwithandroid.ui.theme.MobileTheme
 import org.koin.android.ext.koin.androidLogger
@@ -38,7 +37,6 @@ fun App(
         modules(
             dataSourceModule,
             repositoryModule,
-            stateMachineModule,
             useCaseModule,
         )
     }) {
