+++ mobile/app/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/mobile/AppUiState.kt
@@ -6,9 +6,9 @@ import androidx.compose.runtime.Stable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.remember
-import co.kr.parkjonghun.whatishedoingwithandroid.base.statemachine.StateMachine
-import co.kr.parkjonghun.whatishedoingwithandroid.service.statemachine.sample.SampleAction
-import co.kr.parkjonghun.whatishedoingwithandroid.service.statemachine.sample.SampleState
+import co.kr.parkjonghun.whatishedoingwithandroid.base.usecase.statemachine.StateMachine
+import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.statemachine.sample.SampleAction
+import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.statemachine.sample.SampleState
 import co.kr.parkjonghun.whatishedoingwithandroid.ui.base.Intent
 import co.kr.parkjonghun.whatishedoingwithandroid.ui.base.UiState
 import kotlinx.coroutines.flow.map
