--- mobile/core/domain/base/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/base/statemachine/Action.kt
@@ -1,6 +1,0 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.base.statemachine
-
-/**
- * Catalyst that causes side effects.
- */
-interface Action
