+++ mobile/core/domain/base/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/base/usecase/statemachine/Action.kt
@@ -1,0 +1,6 @@
+package co.kr.parkjonghun.whatishedoingwithandroid.base.usecase.statemachine
+
+/**
+ * Catalyst that causes side effects.
+ */
+interface Action
