+++ mobile/core/domain/base/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/base/usecase/statemachine/State.kt
@@ -1,4 +1,4 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.base.statemachine
+package co.kr.parkjonghun.whatishedoingwithandroid.base.usecase.statemachine
 
 import android.os.Parcelable
