+++ mobile/core/domain/service/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/service/usecase/di/UseCaseModule.kt
@@ -1,18 +1,10 @@
 package co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.di
 
-import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.sample.SampleUseCase
-import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.sample.SampleUseCaseImpl
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
+import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.statemachine.di.stateMachineModule
+import co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.validator.di.validatorModule
 import org.koin.dsl.module
 
 val useCaseModule = module {
-    single<SampleUseCase> {
-        val scope = CoroutineScope(Dispatchers.Main.immediate)
-
-        SampleUseCaseImpl(
-            repository = get(),
-            scope = scope,
-        )
-    }
+    plus(stateMachineModule)
+    plus(validatorModule)
 }
