+++ mobile/core/domain/service/src/main/java/co/kr/parkjonghun/whatishedoingwithandroid/service/usecase/statemachine/sample/SampleState.kt
@@ -1,7 +1,7 @@
-package co.kr.parkjonghun.whatishedoingwithandroid.service.statemachine.sample
+package co.kr.parkjonghun.whatishedoingwithandroid.service.usecase.statemachine.sample
 
-import co.kr.parkjonghun.whatishedoingwithandroid.base.statemachine.ErrorState
-import co.kr.parkjonghun.whatishedoingwithandroid.base.statemachine.State
+import co.kr.parkjonghun.whatishedoingwithandroid.base.usecase.statemachine.ErrorState
+import co.kr.parkjonghun.whatishedoingwithandroid.base.usecase.statemachine.State
 import co.kr.parkjonghun.whatishedoingwithandroid.service.model.sample.SampleToken
 import kotlinx.parcelize.Parcelize
